Reworked class and method placement
